+++ test-suite/golden-tests/arrow-opaque-columns/src.R
@@ -1,0 +1,4 @@
+price <- function(t) {
+  col <- t$column(1)
+  paste(format(as.vector(col), nsmall = 2), col$type$ToString())
+}